#RI-4530 redactor redis-cloud module to be a part of cloud autodiscovery module inside cloud parent module
