+++ redisinsight/api/src/modules/cloud/autodiscovery/dto/add-cloud-database.dto.ts
@@ -1,0 +1,22 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { IsDefined, IsInt, IsNotEmpty } from 'class-validator';
+
+export class AddCloudDatabaseDto {
+  @ApiProperty({
+    description: 'Subscription id',
+    type: Number,
+  })
+  @IsDefined()
+  @IsNotEmpty()
+  @IsInt({ always: true })
+  subscriptionId: number;
+
+  @ApiProperty({
+    description: 'Database id',
+    type: Number,
+  })
+  @IsDefined()
+  @IsNotEmpty()
+  @IsInt({ always: true })
+  databaseId: number;
+}
