+++ redisinsight/api/src/modules/cloud/autodiscovery/dto/index.ts
@@ -1,0 +1,7 @@
+export * from './add-cloud-database.dto';
+export * from './add-cloud-database.response';
+export * from './add-cloud-databases.dto';
+export * from './cloud-auth.dto';
+export * from './get-cloud-subscription-database.dto';
+export * from './get-cloud-subscription-databases.dto';
+export * from './get-cloud-databases.dto';
