+++ redisinsight/api/src/modules/cloud/autodiscovery/models/cloud-subscription.ts
@@ -1,0 +1,48 @@
+import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
+
+export enum CloudSubscriptionStatus {
+  Active = 'active',
+  NotActivated = 'not_activated',
+  Deleting = 'deleting',
+  Pending = 'pending',
+  Error = 'error',
+}
+
+export class CloudSubscription {
+  @ApiProperty({
+    description: 'Subscription id',
+    type: Number,
+  })
+  id: number;
+
+  @ApiProperty({
+    description: 'Subscription name',
+    type: String,
+  })
+  name: string;
+
+  @ApiProperty({
+    description: 'Number of databases in subscription',
+    type: Number,
+  })
+  numberOfDatabases: number;
+
+  @ApiProperty({
+    description: 'Subscription status',
+    enum: CloudSubscriptionStatus,
+    default: CloudSubscriptionStatus.Active,
+  })
+  status: CloudSubscriptionStatus;
+
+  @ApiPropertyOptional({
+    description: 'Subscription provider',
+    type: String,
+  })
+  provider?: string;
+
+  @ApiPropertyOptional({
+    description: 'Subscription region',
+    type: String,
+  })
+  region?: string;
+}
