+++ redisinsight/api/src/modules/cloud/cloud.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { CloudAutodiscoveryModule } from 'src/modules/cloud/autodiscovery/cloud.autodiscovery.module';
+
+@Module({})
+export class CloudModule {
+  static register() {
+    return {
+      module: CloudModule,
+      imports: [CloudAutodiscoveryModule],
+    };
+  }
+}
