--- redisinsight/api/src/modules/redis-enterprise/models/redis-cloud-account.ts
@@ -1,22 +1,0 @@
-export interface IRedisCloudAccount {
-  id: number;
-  name: string;
-  createdTimestamp: string;
-  updatedTimestamp: string;
-  key: IRedisCloudAccountKey;
-}
-
-interface IRedisCloudAccountKey {
-  name: string;
-  accountId: number;
-  accountName: string;
-  allowedSourceIps: string[];
-  createdTimestamp: string;
-  owner: IRedisCloudAccountOwner;
-  httpSourceIp: string;
-}
-
-interface IRedisCloudAccountOwner {
-  name: string;
-  email: string;
-}
