+++ redisinsight/api/src/modules/redis-enterprise/redis-enterprise.analytics.spec.ts
@@ -2,12 +2,9 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { EventEmitter2 } from '@nestjs/event-emitter';
 import { TelemetryEvents } from 'src/constants';
 import {
-  mockRedisCloudDatabaseDto,
-  mockRedisCloudSubscriptionDto,
   mockRedisEnterpriseDatabaseDto,
 } from 'src/__mocks__';
 import { RedisEnterpriseDatabaseStatus } from 'src/modules/redis-enterprise/models/redis-enterprise-database';
-import { RedisCloudSubscriptionStatus } from 'src/modules/redis-enterprise/models/redis-cloud-subscriptions';
 import { InternalServerErrorException } from '@nestjs/common';
 import { RedisEnterpriseAnalytics } from 'src/modules/redis-enterprise/redis-enterprise.analytics';
 
@@ -124,187 +121,6 @@ describe('RedisEnterpriseAnalytics', () => {
 
       expect(sendFailedEventMethod).toHaveBeenCalledWith(
         TelemetryEvents.REClusterDiscoveryFailed,
-        httpException,
-      );
-    });
-  });
-
-  describe('sendGetRECloudSubsSucceedEvent', () => {
-    it('should emit event with active subscriptions', () => {
-      service.sendGetRECloudSubsSucceedEvent([
-        mockRedisCloudSubscriptionDto,
-        mockRedisCloudSubscriptionDto,
-      ]);
-
-      expect(sendEventMethod).toHaveBeenCalledWith(
-        TelemetryEvents.RECloudSubscriptionsDiscoverySucceed,
-        {
-          numberOfActiveSubscriptions: 2,
-          totalNumberOfSubscriptions: 2,
-        },
-      );
-    });
-    it('should emit event with active and not active subscription', () => {
-      service.sendGetRECloudSubsSucceedEvent([
-        {
-          ...mockRedisCloudSubscriptionDto,
-          status: RedisCloudSubscriptionStatus.Error,
-        },
-        mockRedisCloudSubscriptionDto,
-      ]);
-
-      expect(sendEventMethod).toHaveBeenCalledWith(
-        TelemetryEvents.RECloudSubscriptionsDiscoverySucceed,
-        {
-          numberOfActiveSubscriptions: 1,
-          totalNumberOfSubscriptions: 2,
-        },
-      );
-    });
-    it('should emit event without active subscriptions', () => {
-      service.sendGetRECloudSubsSucceedEvent([
-        {
-          ...mockRedisCloudSubscriptionDto,
-          status: RedisCloudSubscriptionStatus.Error,
-        },
-        {
-          ...mockRedisCloudSubscriptionDto,
-          status: RedisCloudSubscriptionStatus.Error,
-        },
-      ]);
-      expect(sendEventMethod).toHaveBeenCalledWith(
-        TelemetryEvents.RECloudSubscriptionsDiscoverySucceed,
-        {
-          numberOfActiveSubscriptions: 0,
-          totalNumberOfSubscriptions: 2,
-        },
-      );
-    });
-    it('should emit GetRECloudSubsSucceedEvent event for empty list', () => {
-      service.sendGetRECloudSubsSucceedEvent([]);
-
-      expect(sendEventMethod).toHaveBeenCalledWith(
-        TelemetryEvents.RECloudSubscriptionsDiscoverySucceed,
-        {
-          numberOfActiveSubscriptions: 0,
-          totalNumberOfSubscriptions: 0,
-        },
-      );
-    });
-    it('should emit GetRECloudSubsSucceedEvent event for undefined input value', () => {
-      service.sendGetRECloudSubsSucceedEvent(undefined);
-
-      expect(sendEventMethod).toHaveBeenCalledWith(
-        TelemetryEvents.RECloudSubscriptionsDiscoverySucceed,
-        {
-          numberOfActiveSubscriptions: 0,
-          totalNumberOfSubscriptions: 0,
-        },
-      );
-    });
-    it('should not throw on error when sending GetRECloudSubsSucceedEvent event', () => {
-      const input: any = {};
-
-      expect(() => service.sendGetRECloudSubsSucceedEvent(input)).not.toThrow();
-      expect(sendEventMethod).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('sendGetRECloudSubsFailedEvent', () => {
-    it('should emit GetRECloudSubsFailedEvent event', () => {
-      service.sendGetRECloudSubsFailedEvent(httpException);
-
-      expect(sendFailedEventMethod).toHaveBeenCalledWith(
-        TelemetryEvents.RECloudSubscriptionsDiscoveryFailed,
-        httpException,
-      );
-    });
-  });
-
-  describe('sendGetRECloudDbsSucceedEvent', () => {
-    it('should emit event with active databases', () => {
-      service.sendGetRECloudDbsSucceedEvent([
-        mockRedisCloudDatabaseDto,
-        mockRedisCloudDatabaseDto,
-      ]);
-
-      expect(sendEventMethod).toHaveBeenCalledWith(
-        TelemetryEvents.RECloudDatabasesDiscoverySucceed,
-        {
-          numberOfActiveDatabases: 2,
-          totalNumberOfDatabases: 2,
-        },
-      );
-    });
-    it('should emit event with active and not active database', () => {
-      service.sendGetRECloudDbsSucceedEvent([
-        {
-          ...mockRedisCloudDatabaseDto,
-          status: RedisEnterpriseDatabaseStatus.Pending,
-        },
-        mockRedisCloudDatabaseDto,
-      ]);
-
-      expect(sendEventMethod).toHaveBeenCalledWith(
-        TelemetryEvents.RECloudDatabasesDiscoverySucceed,
-        {
-          numberOfActiveDatabases: 1,
-          totalNumberOfDatabases: 2,
-        },
-      );
-    });
-    it('should emit event without active databases', () => {
-      service.sendGetRECloudDbsSucceedEvent([
-        {
-          ...mockRedisCloudDatabaseDto,
-          status: RedisEnterpriseDatabaseStatus.Pending,
-        },
-      ]);
-
-      expect(sendEventMethod).toHaveBeenCalledWith(
-        TelemetryEvents.RECloudDatabasesDiscoverySucceed,
-        {
-          numberOfActiveDatabases: 0,
-          totalNumberOfDatabases: 1,
-        },
-      );
-    });
-    it('should emit event for empty list', () => {
-      service.sendGetRECloudDbsSucceedEvent([]);
-
-      expect(sendEventMethod).toHaveBeenCalledWith(
-        TelemetryEvents.RECloudDatabasesDiscoverySucceed,
-        {
-          numberOfActiveDatabases: 0,
-          totalNumberOfDatabases: 0,
-        },
-      );
-    });
-    it('should emit event for undefined input value', () => {
-      service.sendGetRECloudDbsSucceedEvent(undefined);
-
-      expect(sendEventMethod).toHaveBeenCalledWith(
-        TelemetryEvents.RECloudDatabasesDiscoverySucceed,
-        {
-          numberOfActiveDatabases: 0,
-          totalNumberOfDatabases: 0,
-        },
-      );
-    });
-    it('should not throw on error', () => {
-      const input: any = {};
-
-      expect(() => service.sendGetRECloudDbsSucceedEvent(input)).not.toThrow();
-      expect(sendEventMethod).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('sendGetRECloudDbsFailedEvent', () => {
-    it('should emit event', () => {
-      service.sendGetRECloudDbsFailedEvent(httpException);
-
-      expect(sendFailedEventMethod).toHaveBeenCalledWith(
-        TelemetryEvents.RECloudDatabasesDiscoveryFailed,
         httpException,
       );
     });
