--- redisinsight/api/src/modules/redis-enterprise/utils/redis-cloud-converter.ts
@@ -1,5 +1,0 @@
-import { RE_CLOUD_MODULES_NAMES } from 'src/constants';
-
-export function convertRECloudModuleName(name: string): string {
-  return RE_CLOUD_MODULES_NAMES[name] ?? name;
-}
