+++ redisinsight/api/test/api/cloud/POST-cloud-autodiscovery-get_account.test.ts
@@ -1,16 +1,17 @@
 import {
   describe,
-  it,
   deps,
-  validateApiCall,
   requirements,
   generateInvalidDataTestCases,
   validateInvalidDataTestCase,
   Joi,
+  nock, getMainCheckFn,
+  serverConfig,
 } from '../deps';
+import { mockCloudAccountInfo, mockCloudApiAccount } from 'src/__mocks__/cloud-autodiscovery';
 const { request, server, constants } = deps;
 
-const endpoint = () => request(server).post(`/redis-enterprise/cloud/get-account`);
+const endpoint = () => request(server).post(`/cloud/autodiscovery/get-account`);
 
 const dataSchema = Joi.object({
   apiKey: Joi.string().required(),
@@ -29,17 +30,12 @@ const responseSchema = Joi.object().keys({
   ownerEmail: Joi.string().required(),
 }).required();
 
-const mainCheckFn = async (testCase) => {
-  it(testCase.name, async () => {
-    await validateApiCall({
-      endpoint,
-      ...testCase,
-    });
-  });
-};
+const mainCheckFn = getMainCheckFn(endpoint);
 
-describe('POST /redis-enterprise/cloud/get-account', () => {
-  requirements('rte.cloud');
+const nockScope = nock(serverConfig.get('redis_cloud').url);
+
+describe('POST /cloud/autodiscovery/get-account', () => {
+  requirements('rte.serverType=local');
 
   describe('Validation', () => {
     generateInvalidDataTestCases(dataSchema, validInputData).map(
@@ -50,14 +46,49 @@ describe('POST /redis-enterprise/cloud/get-account', () => {
   describe('Common', () => {
     [
       {
+        before: () => {
+          nockScope.get('/')
+            .reply(200, { account: mockCloudApiAccount });
+        },
         name: 'Should get account info',
         data: {
           apiKey: constants.TEST_CLOUD_API_KEY,
           apiSecretKey: constants.TEST_CLOUD_API_SECRET_KEY,
         },
         responseSchema,
+        responseBody: mockCloudAccountInfo,
       },
       {
+        before: () => {
+          nockScope.get('/')
+            .reply(403, {
+              response: {
+                status: 403,
+                data: { message: 'Unauthorized for this action' },
+              }
+            });
+        },
+        name: 'Should throw Forbidden error when api returned unauthorized error',
+        data: {
+          apiKey: constants.TEST_CLOUD_API_KEY,
+          apiSecretKey: constants.TEST_CLOUD_API_SECRET_KEY,
+        },
+        statusCode: 403,
+        responseBody: {
+          statusCode: 403,
+          error: 'Forbidden',
+        },
+      },
+      {
+        before: () => {
+          nockScope.get('/')
+            .reply(401, {
+              response: {
+                status: 401,
+                data: '',
+              }
+            });
+        },
         name: 'Should throw Forbidden error when api key is incorrect',
         data: {
           apiKey: 'wrong-api-key',
@@ -68,20 +99,6 @@ describe('POST /redis-enterprise/cloud/get-account', () => {
           statusCode: 403,
           error: 'Forbidden',
         },
-
-      },
-      {
-        name: 'Should throw Forbidden error when api secret key is incorrect',
-        data: {
-          apiKey: constants.TEST_CLOUD_API_KEY,
-          apiSecretKey: 'wrong-api-secret-key',
-        },
-        statusCode: 403,
-        responseBody: {
-          statusCode: 403,
-          error: 'Forbidden',
-        },
-
       },
     ].map(mainCheckFn);
   });
