+++ redisinsight/api/test/api/deps.ts
@@ -12,6 +12,12 @@ import { testEnv } from '../helpers/test';
 import * as redis from '../helpers/redis';
 import { initCloudDatabase } from '../helpers/cloud';
 
+// Just dummy jest module implementation to be able to use common mocked models in UTests and ITests
+global['jest'] = {
+  // @ts-ignore
+  fn: () => {}
+};
+
 /**
  * Initialize dependencies
  */
