+++ redisinsight/api/test/helpers/constants.ts
@@ -103,10 +103,10 @@ export const constants = {
   // cloud
   TEST_CLOUD_RTE: process.env.TEST_CLOUD_RTE,
   TEST_CLOUD_API: process.env.REDIS_CLOUD_URL || process.env.TEST_CLOUD_API || 'https://api.qa.redislabs.com/v1',
-  TEST_CLOUD_API_KEY: process.env.TEST_CLOUD_API_KEY,
-  TEST_CLOUD_API_SECRET_KEY: process.env.TEST_CLOUD_API_SECRET_KEY,
+  TEST_CLOUD_API_KEY: process.env.TEST_CLOUD_API_KEY || 'TEST_CLOUD_API_KEY',
+  TEST_CLOUD_API_SECRET_KEY: process.env.TEST_CLOUD_API_SECRET_KEY || 'TEST_CLOUD_API_SECRET_KEY',
   TEST_CLOUD_SUBSCRIPTION_NAME: process.env.TEST_CLOUD_SUBSCRIPTION_NAME || 'ITests',
-  TEST_CLOUD_SUBSCRIPTION_ID: process.env.TEST_CLOUD_SUBSCRIPTION_ID,
+  TEST_CLOUD_SUBSCRIPTION_ID: process.env.TEST_CLOUD_SUBSCRIPTION_ID || 1,
   TEST_CLOUD_DATABASE_NAME: process.env.TEST_CLOUD_DATABASE_NAME || 'ITests-db',
 
   STANDALONE: 'STANDALONE',
